move framework ESlint settings 100% to ./.eslintrc.js
